+++ diventi/accounts/translation.py
@@ -1,7 +1,11 @@
 from modeltranslation.translator import register, TranslationOptions
 
-from .models import DiventiAvatar, DiventiCover, Achievement
+from .models import DiventiAvatar, DiventiCover, Achievement, DiventiUser
 
+
+@register(DiventiUser)
+class DiventiAvatarTranslationOptions(TranslationOptions):
+    fields = ('bio',)    
 
 @register(DiventiAvatar)
 class DiventiAvatarTranslationOptions(TranslationOptions):